Move files into app_manager.app_schemas
